Refactor importFile to getFileStream
This change is because rather than open a reading large file, we need to use streams.
To allow decoupling this will be done in multiple steps, the getFileStream needs only to validate that we can get the file.
Returning it to another method allows us to extend behaviour quickly and easily.